Create lanes_basic
Basic lane implementation, each lane currently acts on its own.
Next step is to introduce "Smart" Movement. 
Cars will look 'left' and 'right' (i.e lane above and lane below) If the lanes adjacent allow for the car to increase it's velocity (i.e distance to car infront is greater than current lane) the car will switch lanes. We should iterate backwards, i.e the car at the front of the pack makes it's decision first, the other cars react to it. The code atm is kind of hard to read will clean up later, just wanted to get it working first. It works with the same car class currently so that needs no changes, I think the 'Smart' Movement function needs to be done in the Road class and not the car class, since the Car class doesn't have lane information.