Create lanes_basic.py
Tried to use the new distance_to_next function, sometimes it works sometimes,sometimes it doesnt. It seems cars disappear at the end only if they are fast enough. If a car gets initialized at the last spot with speed zero, then it doesn't have enough speed to vanish and thinks theres a car at the end or it checks to see the spot infront of it and it doesn't know theres no more road so . and stays there forever.  Sometimes there's a crash at the end of the road because of this, and the congestion is never resolved I have added screenshots in the repository